Minor refactory to fix the build
